+++ src/Action/PortalNodeStorage/PortalNodeStorageClear.php
@@ -12,14 +12,19 @@ use Heptacom\HeptaConnect\Storage\Base\Exception\DeleteException;
 use Heptacom\HeptaConnect\Storage\Base\Exception\UnsupportedStorageKeyException;
 use Heptacom\HeptaConnect\Storage\Base\PreviewPortalNodeKey;
 use Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey\PortalNodeStorageKey;
-use Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Query\QueryBuilder;
+use Heptacom\HeptaConnect\Storage\ShopwareDal\Support\Query\QueryFactory;
 
 class PortalNodeStorageClear implements PortalNodeStorageClearActionInterface
 {
+    public const CLEAR_QUERY = '1087e0dc-07fe-48d7-903c-9353167c3e89';
+
+    private QueryFactory $queryFactory;
+
     private Connection $connection;
 
-    public function __construct(Connection $connection)
+    public function __construct(QueryFactory $queryFactory, Connection $connection)
     {
+        $this->queryFactory = $queryFactory;
         $this->connection = $connection;
     }
 
@@ -35,7 +40,7 @@ class PortalNodeStorageClear implements PortalNodeStorageClearActionInterface
             throw new UnsupportedStorageKeyException(\get_class($portalNodeKey));
         }
 
-        $deleteBuilder = new QueryBuilder($this->connection);
+        $deleteBuilder = $this->queryFactory->createBuilder(self::CLEAR_QUERY);
         $deleteBuilder
             ->delete('heptaconnect_portal_node_storage')
             ->andWhere($deleteBuilder->expr()->eq('portal_node_id', ':portal_node_id'))